refactor: update restaurant data handling and remove Geoapify integration
- Remove the duplicated code in restaurants.ts and geoapify.ts
- Update restaurant fetching function name for clarity
- Adjust components to utilize the new restaurant data structure
