+++ src/app/groups/page.tsx
@@ -1,17 +1,14 @@
+"use client";
+
 import { Plus } from "lucide-react";
 import Link from "next/link";
-import { redirect } from "next/navigation";
 
-import { groups } from "@/data/groups";
-import checkUser from "@/hooks/check-user";
+import { useGroups } from "@/contexts/group-context";
 import { GroupCard } from "~/groups/card";
 import { Button } from "~/ui/button";
 
-export default async function Groups() {
-  const loggedIn = await checkUser();
-  if (!loggedIn) {
-    redirect("/login");
-  }
+export default function Groups() {
+  const { groups } = useGroups();
 
   return (
     <>
